# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MarocExplore API",
        "description": "API documentation for the MarocExplore project, providing endpoints for itinerary management, destinations, and user authentication.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://marocexplore.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/itineraries/{itinerary}/destinations": {
            "post": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Add a destination to an itinerary",
                "operationId": "4a466508c0409e5c28fd3dc59ad69ab9",
                "parameters": [
                    {
                        "name": "itinerary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "address"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Sahara Desert"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Merzouga"
                                    },
                                    "places": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "activities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "dishes": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Destination added"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/destinations/{destination}": {
            "put": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Update a destination",
                "operationId": "f86e483d05239cb06ede4100353802aa",
                "parameters": [
                    {
                        "name": "destination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "places": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "activities": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "dishes": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Destination updated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Destinations"
                ],
                "summary": "Delete a destination",
                "operationId": "374d85b3725e458bfc00388f4b85343e",
                "parameters": [
                    {
                        "name": "destination",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Destination deleted"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}